Title: 만나는 그 순간, Time: 17ms, Memory: 0MB, Status: Runtime Error - CodeTree

+++ 240413/만나는 그 순간/the-moment-we-meet.cpp
@@ -13,7 +13,7 @@ int main() {
     int t;
     //A이동
     int a_idx = 1;
-    int total_time = 0;
+    unsigned long long total_time = 0;
     for(int i = 1; i <= n; i++) {
         cin >> d >> t;
         total_time += t;
